+++ src/main/scala/edu/ie3/util/scala/quantities/DefaultQuantities.scala
@@ -7,6 +7,7 @@
 package edu.ie3.util.scala.quantities
 
 import squants.energy.{KilowattHours, Kilowatts, MegawattHours, Megawatts}
+import squants.radio.{Irradiance, WattsPerSquareMeter}
 import squants.thermal.Celsius
 import squants.{Dimensionless, Each, Energy, Power, Temperature}
 
@@ -22,6 +23,8 @@ object DefaultQuantities {
 
   val zeroKWh: Energy = KilowattHours(0d)
   val zeroMWh: Energy = MegawattHours(0d)
+
+  val zeroEurPerKWh: EnergyPrice = EuroPerKilowattHour(0d)
 
   val zeroPU: Dimensionless = Each(0d)
   val onePU: Dimensionless = Each(1d)