+++ src/main/scala/edu/ie3/util/scala/quantities/SquantsUtils.scala
@@ -8,23 +8,25 @@ package edu.ie3.util.scala.quantities
 
 import squants.electro.{ElectricPotential, Volts}
 import squants.energy.Energy
+import squants.radio.Irradiance
 import squants.space.{CubicMeters, Volume}
 import squants.thermal.ThermalCapacity
-import squants.{Dimensionless, Each}
+import squants.time.Hours
+import squants.{Dimensionless, Each, Time}
 
 object SquantsUtils {
-  implicit class RichEnergy(energy: Energy) {
+  extension (energy: Energy) {
     def calcVolume(that: EnergyDensity): Volume = CubicMeters(
       energy.toKilowattHours / that.toKilowattHoursPerCubicMeter
     )
   }
-  implicit class RichPower(power: squants.Power) {
+  extension (power: squants.Power) {
     def /(that: ReactivePower): Dimensionless = Each(
       power.toWatts / that.toVars
     )
   }
 
-  implicit class RichElectricPotential(
+  extension (
       electricPotential: ElectricPotential
   ) {
     def multiplyWithDimensionless(
@@ -35,12 +37,20 @@ object SquantsUtils {
 
   }
 
-  implicit class RichThermalCapacity(
+  extension (
       thermalCapacity: ThermalCapacity
   ) {
     def toWattHoursPerKelvin: Double =
-      this.thermalCapacity.toJoulesPerKelvin / 3600
+      thermalCapacity.toJoulesPerKelvin / 3600
     def toWattSecondsPerKelvin: Double =
-      this.thermalCapacity.toJoulesPerKelvin // Joule == Ws
+      thermalCapacity.toJoulesPerKelvin // Joule == Ws
+  }
+
+  extension (
+      irradiance: Irradiance
+  ) {
+    def *(that: Time): Irradiation = WattHoursPerSquareMeter(
+      irradiance.toWattsPerSquareMeter * that.toSeconds / Hours(1).toSeconds
+    )
   }
 }